Progression on UI. Moved "width" and "height" back to unsigned int.
git-svn-id: file:///home/michael/svn/yuni/trunk@1239 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/gfx/surface/cairo.h
@@ -26,7 +26,7 @@ namespace Surface
 		virtual void release();
 
 		//! Resize the surface
-		virtual void resize(size_t width, size_t height);
+		virtual void resize(unsigned int width, unsigned int height);
 
 		//! Choose a color to clear the screen with
 		virtual void clearColor(Color::RGB<uint8>& newColor);
